style: PC환경 선택일 지정 UI 좌/우 1+1 형식으로 변경
